Implement logging using up-to-dated code
This modified implementation is used in the up-to-date code. and will just record the critical process such as the cleaning start and end, all the kinds of warnings, etc. It will not record the button creation and any other debug information but record if there is an error.

Tested all the stuff, works fine.